refactor : 사다리게임 step2 CreatePlayer클래스 삭제관련 리팩토링 객체의 생성 책임을 명확하게 하기 위해 Players 객체로 생성 관련 메소드를 전부 인관. CreatePlayer클래스 삭제.
