+++ src/test/java/ladderGame/step2/model/PlayersTest.java
@@ -2,7 +2,6 @@ package ladderGame.step2.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import ladderGame.step2.service.CreatePlayer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +10,7 @@ class PlayersTest {
   @DisplayName("사용자명을 쉼표단위로 여러개 입력하여 사용자를 생성.")
   @Test
   void createPlayersTest() {
-    Players players = new Players(CreatePlayer.createUser("user1,user2"));
+    Players players = new Players(Players.createUser("user1,user2"));
     assertThat(players.count()).isEqualTo(2);
   }
 }
