+++ src/test/java/ladderGame/step2/service/CreatePlayerTest.java
@@ -3,6 +3,7 @@ package ladderGame.step2.service;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import ladderGame.step2.model.Player;
+import ladderGame.step2.model.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,14 +13,14 @@ class CreatePlayerTest {
   @Test
   void createPlayers() {
 
-    assertThat(CreatePlayer.createUser("user1,user2").size()).isEqualTo(2);
-    assertThat(CreatePlayer.createUser("user1,user2")).contains(new Player("user1"));
-    assertThat(CreatePlayer.createUser("user1,user2")).contains(new Player("user2"));
+    assertThat(Players.createUser("user1,user2").size()).isEqualTo(2);
+    assertThat(Players.createUser("user1,user2")).contains(new Player("user1"));
+    assertThat(Players.createUser("user1,user2")).contains(new Player("user2"));
   }
 
   @DisplayName("중복되는 이름을 자동으로 제거하고 생성.")
   @Test
   void distinctDuplicateName() {
-    assertThat(CreatePlayer.createUser("user3,user3").size()).isEqualTo(1);
+    assertThat(Players.createUser("user3,user3").size()).isEqualTo(1);
   }
 }
